little more refactor, splitting out components
